requirements/development.txt
-r production.txt

flake8
requests
ipython
beautifulsoup4
django-debug-toolbar
django-extensions
